Don't interpret ampersands as shortcut keys
For the labels in the details block, substitute ampersands for
double ampersands so Qt framework knows I don't want a
shortcut key here.

I did this by creating an ExplicitPushButton class so this is
transparent to the calling code. I'm not in love with the
call-super to QPushButton in setText() but it seems like the
best way to do it with the options that I have.